# openapi.yml
openapi: 3.1.0
info:
  title: Система бронирования мест в кафе
  version: 0.0.4
servers:
  - url: http://localhost:8000/
    description: Локальный сервер
...
paths:
...
  /users:
    get:
      tags:
        - Пользователи
      summary: Получение списка пользователей
      description: Возвращает информацию о всех пользователях. Только для администраторов или менеджеров
      operationId: get_all_user_users_get
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserInfo'
                type: array
                title: Response Get All User Users Get
        '401':
          description: Неавторизированный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
      security:
        - OAuth2PasswordBearer: []
    post:
      tags:
        - Пользователи
      summary: Регистрация нового пользователя
      description: Создает нового пользователя с указанными данными. <br><br><b>Обязательные поля:</b> <ul><li>username</li><li>password</li><li>email или phone</li></ul>
      operationId: create_user_users_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '400':
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '422':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
  /users/{user_id}:
    get:
      tags:
        - Пользователи
      summary: Получение информации о пользователе по его ID
      description: Возвращает информацию о пользователе по его ID. Только для администраторов или менеджеров
      operationId: get_user_by_id_users__user_id__get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            title: User Id
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          description: Неавторизированный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '404':
          description: Данные не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '422':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
    patch:
      tags:
        - Пользователи
      summary: Обновление информации о пользователе по его ID
      description: Возвращает обновленную информацию о пользователе по его ID. Только для администраторов или менеджеров
      operationId: path_user_by_id_users__user_id__patch
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '400':
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '401':
          description: Неавторизированный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '404':
          description: Данные не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '422':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
  /users/me:
    get:
      tags:
        - Пользователи
      summary: Получение информации о текущем пользователе
      description: Возвращает информацию о текущем пользователе. Только для авторизированных пользователей
      operationId: get_me_users_me_get
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          description: Неавторизированный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
      security:
        - OAuth2PasswordBearer: []
    patch:
      tags:
        - Пользователи
      summary: Обновление информации о текущем пользователе
      description: Возвращает обновленную информацию о пользователе. Только для авторизированных пользователей
      operationId: path_me_users_me_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
        required: true
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '400':
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '401':
          description: Неавторизированный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '422':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
      security:
        - OAuth2PasswordBearer: []
...
components:
  schemas:
...
    CustomError:
      properties:
        code:
          type: integer
          title: Code
        message:
          type: string
          title: Message
      type: object
      required:
        - code
        - message
      title: CustomError
...
    UserCreate:
      properties:
        username:
          type: string
          title: Username
        email:
          type: string
          title: Email
        phone:
          type: string
          title: Phone
        tg_id:
          type: string
          title: Tg Id
        password:
          type: string
          title: Password
      type: object
      required:
        - username
        - password
      title: UserCreate
    UserInfo:
      properties:
        id:
          type: integer
          title: Id
        username:
          type: string
          title: Username
        email:
          type: string
          title: Email
        phone:
          type: string
          title: Phone
        tg_id:
          type: string
          title: Tg Id
        role:
          $ref: '#/components/schemas/UserRole'
        is_active:
          type: boolean
          title: Is Active
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      title: UserInfo
    UserRole:
      type: integer
      oneOf:
        - title: USER
          const: 0
          description: Пользователь
        - title: MANAGER
          const: 1
          description: Менеджер
        - title: ADMIN
          const: 2
          description: Администратор
      title: UserRole
...
    UserUpdate:
      properties:
        username:
          anyOf:
            - type: string
            - type: 'null'
          title: Username
        email:
          anyOf:
            - type: string
            - type: 'null'
          title: Email
        phone:
          anyOf:
            - type: string
            - type: 'null'
          title: Phone
        tg_id:
          anyOf:
            - type: string
            - type: 'null'
          title: Tg Id
        role:
          anyOf:
            - $ref: '#/components/schemas/UserRole'
            - type: 'null'
        password:
          anyOf:
            - type: string
            - type: 'null'
          title: Password
        is_active:
          anyOf:
            - type: boolean
            - type: 'null'
          title: Is Active
      type: object
      title: UserUpdate
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /auth/login